refactor: move all `address` files from `no-meteor` to `node-app`
