+++ imports/node-app/plugins/address/queries/index.js
@@ -1,0 +1,7 @@
+import addressValidation from "./addressValidation.js";
+import addressValidationServices from "./addressValidationServices.js";
+
+export default {
+  addressValidation,
+  addressValidationServices
+};
