+++ imports/node-app/plugins/address/resolvers/index.js
@@ -1,0 +1,5 @@
+import Query from "./Query/index.js";
+
+export default {
+  Query
+};
