+++ imports/node-app/registerPlugins.js
@@ -1,9 +1,10 @@
 /* node-app imports */
 import registerSystemInfoPlugin from "./plugins/system-info/index.js";
+import registerAddressPlugin from "./plugins/address/index.js";
+
 
 /* meteor-app imports */
 import registerAccountsPlugin from "/imports/plugins/core/accounts/server/no-meteor/register";
-import registerAddressPlugin from "/imports/plugins/core/address/server/no-meteor/register";
 import registerCartPlugin from "/imports/plugins/core/cart/server/no-meteor/register";
 import registerCatalogPlugin from "/imports/plugins/core/catalog/server/no-meteor/register";
 import registerCheckoutPlugin from "/imports/plugins/core/checkout/server/no-meteor/register";
