--- imports/plugins/core/address/server/no-meteor/resolvers/Query/index.js
@@ -1,7 +1,0 @@
-import addressValidation from "./addressValidation";
-import addressValidationServices from "./addressValidationServices";
-
-export default {
-  addressValidation,
-  addressValidationServices
-};
